Create srtScript
Recently I needed to hard-code (burn in) some subtitles onto a video file. I didn't find any easy way to do this, the very nice HandBrake software that I tried could only embed the SRT file into the mp4 file which was not what I wanted. VirtualDub could do something similar but only with SSA subtitles and could only produce AVI-files. I decided to do it in AE.

The below script does just that: It creates keyframes on a subtitle text layer in AE which can be overlayed to your video and rendered to a file. I've tried it with AE CS4 and CS5 on Windows 7 64bit

Usage:
* Open AE and create your project. Import the video track you want subtitled.
* Add a new text layer and set the font and other attributes you want changed. Note that you can't do this after the importing is finished.
* File->Scripts->Open script editor from the menu.
* Paste the below script into the editor and hit run. Wait a while, this script is SLOW (yeah I know it's weird but apparently keyframes take a long time to create from a script)

Alternatively save the script to a file in the Adobe scripts folder and it will appear in the AE Scripts sub-menu. Good night and good luck!